requirements.txt
fastapi==0.129.0
uvicorn[standard]==0.41.0
motor==3.7.1
pydantic[email]==2.12.5
python-dotenv==1.2.1
pymongo==4.16.0
email-validator==2.3.0
annotated-types==0.7.0
anyio==4.12.1
starlette==0.52.1
typing_extensions==4.15.0
